CG-0MP0B5190004IVC9: extract MainStreetAnimator and delegate transfer/HUD animations
